# api/openapi.yaml
openapi: 3.1.0
info:
  title: ShellOrchestra API
  version: 0.1.0
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /servers:
    get:
      operationId: listServers
      responses:
        '200':
          description: Servers
          content:
            application/json:
              schema:
                type: object
                required: [servers]
                properties:
                  servers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Server'
    post:
      operationId: createServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerInput'
      responses:
        '201':
          description: Created server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
...
  /servers/{serverId}:
    put:
      operationId: updateServer
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerInput'
      responses:
        '200':
          description: Updated server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
    delete:
      operationId: deleteServer
      parameters:
        - $ref: '#/components/parameters/ServerId'
      responses:
        '204':
          description: Deleted server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: shellorchestra_access
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ServerInput:
      type: object
      required: [name, host, port, username]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        host:
          type: string
          minLength: 1
          maxLength: 255
        port:
          type: integer
          minimum: 1
          maximum: 65535
        username:
          type: string
          minLength: 1
          maxLength: 120
        connection_mode:
          type: string
          enum: [direct, chained]
          default: direct
        jump_server_id:
          type: string
          nullable: true
        auth_method:
          type: string
          enum: [ca, classic, custom_key, local_protected_key]
          default: ca
        ssh_key_id:
          type: string
          nullable: true
        shell_hint:
          type: string
          enum: [auto, posix, bash, zsh, powershell]
        os_hint:
          type: string
        distro_hint:
          type: string
        detected_shell:
          type: string
        detected_os:
          type: string
        detected_distro:
          type: string
        detected_admin_rights:
          type: string
        detected_hostname:
          type: string
        detected_platform:
          type: string
        detected_platform_os:
          type: string
        detected_platform_arch:
          type: string
        detected_kernel_version:
          type: string
        detected_package_manager:
          type: string
        detected_ssh_max_sessions:
          type: integer
          minimum: 0
        detected_pve_host:
          type: boolean
        detected_docker_host:
          type: boolean
        detected_apps:
          type: object
          additionalProperties:
            type: boolean
        override_shell:
          type: string
          nullable: true
        override_os:
          type: string
          nullable: true
        override_distro:
          type: string
          nullable: true
        override_admin_rights:
          type: string
          nullable: true
        host_key:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    Server:
      allOf:
        - $ref: '#/components/schemas/ServerInput'
        - type: object
          required: [id, connection_mode, auth_method, shell_hint, os_hint, distro_hint, detected_shell, detected_os, detected_distro, detected_admin_rights, detected_hostname, detected_platform, detected_platform_os, detected_platform_arch, detected_kernel_version, detected_package_manager, detected_ssh_max_sessions, detected_pve_host, detected_docker_host, detected_apps, created_at, updated_at]
          properties:
            id:
              type: string
            shell_hint:
              type: string
              enum: [auto, posix, bash, zsh, powershell]
              description: Backend-detected shell family. Client input is ignored.
            os_hint:
              type: string
              description: Backend-detected operating system family.
            distro_hint:
              type: string
              description: Backend-detected distribution identifier when available.
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...